bk: 3f15a7e2gL2W2M_zvzn_Z2KduCwqpA

+++ src/clone.c
@@ -154,6 +154,7 @@ clone(char **av, opts opts, remote *r, char *local, char **envVar)
 {
 	char	*p, buf[MAXPATH];
 	int	gzip, rc = 1;
+	char	*lic;
 
 	gzip = r->port ? opts.gzip : 0;
 	if (local && exists(local) && !emptyDir(local)) {
@@ -196,6 +197,12 @@ clone(char **av, opts opts, remote *r, char *local, char **envVar)
 		disconnect(r, 2);
 		goto done;
 	}
+	if ((lic = getenv("BKD_LICTYPE")) && !licenseAcceptOne(1, lic)) {
+		fprintf(stderr, "clone: failed to accept license '%s'\n", getenv("BKD_LICTYPE"));
+		disconnect(r, 2);
+		goto done;
+	}
+
 	unless (opts.quiet) {
 		remote	*l = remote_parse(local, 0);
 
@@ -259,10 +266,17 @@ private	int
 clone2(opts opts, remote *r)
 {
 	char	*p;
-	char	*checkfiles = bktmp(0, "clone2");
+	char	*checkfiles;
 	FILE	*f;
 	int	rc;
 
+	unless (licenseAccept(1)) {
+		fprintf(stderr, "clone failed license accept check\n");
+		unlink("SCCS/s.ChangeSet");
+		return (-1);
+	}
+
+	checkfiles = bktmp(0, "clone2");
 	f = fopen(checkfiles, "w");
 	assert(f);
 	sccs_rmUncommitted(opts.quiet, f);
@@ -602,7 +616,7 @@ out:		chdir(from);
 	chdir(from);
 	repository_rdunlock(0);
 	fromid = repo_id();
-	chdir(dest);
+	if (chdir(dest)) goto out;
 	rmdir("RESYNC");		/* undo wants it gone */
 	if (clone2(opts, r)) {
 		mkdir("RESYNC", 0777);
